==> Jupyter Notebooks/1 - Get fasta peptides.ipynb <==
### Total theoretical peptides (FASTA)    

from google.colab import drive
drive.mount('/content/drive')

!python --version

# import libraries
import numpy as np
import pandas as pd
import re
!pip install Bio
from Bio import SeqIO
import csv

### Import df_ones_sorted_PTEN file

### File obtained from Ahmad - 'https://github.com/i-am-robot3930/Tryptic-Cleavage'
fasta = pd.read_table('/content/drive/MyDrive/Colab Notebooks/peptide ml detection/data/df_ones_sorted_PTEN.csv', header=None)
fasta.head()

### changes made to column names, from protein id to protein and peptides
### columns with missed cleavage value of '1' were filtered and saved as tsv file
df_ones_sorted_ones = pd.read_table('/content/df_ones_sorted_PTEN final.tsv')
df_ones_sorted_ones.head()

### file renamed
fasta_proteins_new = pd.read_table('/content/fasta_proteins.tsv')
fasta_proteins_new.head()

### Add sequences column to calculate their length and add them into 'sequence_length' column (i.e, sequence length for each protein should be calculated and those values are to be added to a 'sequence_length' column.
### This is done to calculate protein quantitation metrics later

import pandas as pd

# list_of_sequence = ["" for i in range (len(df_ones_sorted_ones. index))]
# list_of_sequence_length = [0 for i in range (len(df_ones_sorted_ones. index))]
list_of_sequence=[]
list_of_sequence_length=[]
# df_ones_sorted_ones = df_ones_sorted_ones.assign(sequence = list_of_sequence)
# df_ones_sorted_ones = df_ones_sorted_ones.assign(sequence_length = list_of_sequence_length)
for index, row in df_ones_sorted_ones.iterrows():
  new_row = fasta_proteins_new.loc[fasta_proteins_new["Protein"] == row["Protein"]]
  if(new_row.size == 0):
    list_of_sequence.append("")
    list_of_sequence_length.append(0)
    continue;
  list_of_sequence.append(new_row["Sequence"].values[0])
  list_of_sequence_length.append(len(new_row["Sequence"].values[0]))
  # row["sequence_length"] = len(new_row["Sequence"].values[0])
  # row["sequence"] = new_row["Sequence"].values[0]
  print(len(new_row["Sequence"].values[0]))
  # break
df_ones_sorted_ones = df_ones_sorted_ones.assign(sequence = list_of_sequence)
df_ones_sorted_ones = df_ones_sorted_ones.assign(sequence_length = list_of_sequence_length)
df_ones_sorted_ones.head()
df_ones_sorted_ones.to_csv('newfile.tsv', sep = '\t')

fasta_proteins = pd.read_table('/content/newfile.tsv')
fasta_proteins.head()

#### To verify if the missed cleavages in the written code works correctly on FASTA sequences, check the following

### KKKK+ should not exist in peptide sequences
len(fasta_peptides[fasta_peptides['Peptide'].str.contains("KKKK", na=False)])

### should not allow peptides with "KKK"
len(fasta_peptides[fasta_peptides['Peptide'].str.contains("KKK", na=False)])

### should not allow peptides with "KK"
len(fasta_peptides[fasta_peptides['Peptide'].str.contains("KK", na=False)])

### RRRR+ should not exist in peptide sequences
len(fasta_peptides[fasta_peptides['Peptide'].str.contains("RRRR", na=False)])

### should not allow peptides with "RRR"
len(fasta_peptides[fasta_peptides['Peptide'].str.contains("RRR", na=False)])

Desired file display

### file saved as fasta_peptides

fasta_peptides = pd.read_table('/content/drive/MyDrive/Colab Notebooks/peptide ml detection/data/fasta_peptides.tsv')
fasta_peptides.head()
