==> Jupyter Notebooks/2- Calculate detected peptides.ipynb <==
# Calculate detected peptides

## Notebook summary:
1. Import MaxQuant output (evidence.txt) file.
2. Clean dataset.
    - Check for and remove contaminants and 'reverse' sequences by MaxQuant and empty protein IDs.
    - Keep only peptides with 1 missed cleavages.
3. Reformat dataset columns for ML training and export reformatted peptide dataset as a TSV file.


from google.colab import drive
drive.mount('/content/drive')

# import libraries
import numpy as np
import pandas as pd
import re
import csv

# set display options
#pd.set_option("display.max_rows", None, "display.max_columns", None)

### 1. Import MaxQuant output (evidence.txt) file

# load dataset
evidence = pd.read_table('/content/drive/MyDrive/Colab Notebooks/peptide ml detection/data/evidence.txt')
evidence.head()

#### Explore dataset

evidence.shape

evidence.columns

# get summary
evidence.describe()

#### Verification: check number of proteins (unique rows) and peptides

evidence['Proteins'].describe()

evidence['Sequence'].describe()

# there are more sequences than Proteins - maybe there are some missing proteins? Check for this
print(evidence['Proteins'].isnull().sum()) # after a manual check these seem to be all reverse sequences

# check missed cleavages
evidence['Missed cleavages'].describe(include=['category'])

### 2. Clean dataset

#### Remove contaminants, 'reverse' sequences and empty protein IDs

# check number of contaminant peptides
print(len(evidence.loc[evidence['Potential contaminant'] == '+']))

# remove contaminant peptides
evidence_cleaned = evidence.loc[evidence['Potential contaminant'] != '+']
evidence_cleaned.shape

# check number of contaminant peptides after removing
print(len(evidence_cleaned.loc[evidence_cleaned['Potential contaminant'] == '+']))

# check number of 'reverse' sequences before removing contaminants
print(len(evidence.loc[evidence['Reverse'] == '+']))

# check number of 'reverse' sequences after removing contaminants
print(len(evidence_cleaned.loc[evidence_cleaned['Reverse'] == '+']))

# remove 'reverse' sequences
evidence_cleaned = evidence_cleaned.loc[evidence_cleaned['Reverse'] != '+']
evidence_cleaned.shape

# check number of 'reverse' sequences after removing them
print(len(evidence_cleaned.loc[evidence_cleaned['Reverse'] == '+']))

# check for number of proteins
evidence_cleaned['Leading razor protein'].describe()

# check for number of peptides
evidence_cleaned['Sequence'].describe()

#### Keep only peptides with 0 missed cleavages.

# check number of 0 missed cleavages
print(len(evidence_cleaned.loc[evidence_cleaned['Missed cleavages'] == 0]))

# check number of 1 missed cleavages
print(len(evidence_cleaned.loc[evidence_cleaned['Missed cleavages'] == 1]))

# check number of 2 missed cleavages
print(len(evidence_cleaned.loc[evidence_cleaned['Missed cleavages'] == 2]))

# check number of 3 missed cleavages (should be 0)
print(len(evidence.loc[evidence['Missed cleavages'] == 3]))

# keep only the peptides with 1 missed cleavages
evidence_cleaned = evidence_cleaned.loc[(evidence_cleaned['Missed cleavages'] == 1)]

# check for any missed cleavages
evidence_cleaned['Missed cleavages'].describe()

# check number of missed cleavages after removing them
print(len(evidence_cleaned.loc[evidence_cleaned['Missed cleavages'] != 1]))

evidence_cleaned.shape

#### Verification: check total number of missing values for each feature to determine any further rows for removal

print(evidence_cleaned.isnull().sum())
# No missing proteins or sequences, so no need to remove any further rows. Features that do have missing data are expected.

#### Verification: check for any potential co-eluting contaminants and reverse sequences in 'Proteins' and 'leading razor protein' columns

# check for any potential co-eluting contaminants in 'Proteins' column
len(evidence_cleaned[evidence_cleaned['Proteins'].str.contains("CON__", na=False)])

# remove co-eluting contaminants in 'Proteins' column
evidence_cleaned = evidence_cleaned[~evidence_cleaned['Proteins'].str.contains("CON__", na=False)]
evidence_cleaned.shape

# verify if all contaminants are removed in 'Proteins' column
len(evidence_cleaned[evidence_cleaned["Proteins"].str.contains("CON__")])

# check for any co-eluting contaminants in 'Leading razor protein' column
len(evidence_cleaned[evidence_cleaned['Leading razor protein'].str.contains("CON__", na=False)])

# check for any co-eluting reverse sequences in 'Proteins' column
len(evidence_cleaned[evidence_cleaned['Proteins'].str.contains("REV__", na=False)])

# check for any co-eluting reverse sequences in 'Leading razor protein' column
len(evidence_cleaned[evidence_cleaned['Leading razor protein'].str.contains("REV__", na=False)])

evidence_cleaned.shape

# check how many identified peptides map to more than one protein (ambiguous protein groups)
len(evidence_cleaned[evidence_cleaned['Proteins'].str.contains(";")])

# keep only rows that don't contain multiple proteins in "Proteins" column
evidence_cleaned = evidence_cleaned[~evidence_cleaned['Proteins'].str.contains(";", na=False)]
evidence_cleaned.shape

# verify these protein groups were removed
len(evidence_cleaned[evidence_cleaned['Proteins'].str.contains(";")])

# export cleaned dataset as TSV
evidence_cleaned.to_csv("/content/drive/MyDrive/Colab Notebooks/peptide ml detection/data/evidence_cleaned.tsv", sep='\t', index=False)

### 3. Reformat dataset columns for ML training

# extract protein and peptide sequence from evidence_cleaned
# make the "leading razor groups" column as the protein column
detected_peptides = evidence_cleaned[['Proteins', 'Sequence', 'PEP']]
detected_peptides.head()

#### Verification: final check on dimensions after formatting, and if expected number of proteins and peptides are present

# check dimensions
detected_peptides.shape

# rename columns
detected_peptides = detected_peptides.rename(columns={"Proteins": "Protein", "Sequence": "Peptide"})
detected_peptides.head()

### 4. Export reformatted peptide dataset as TSV

detected_peptides.to_csv("/content/drive/MyDrive/Colab Notebooks/peptide ml detection/data/detected_peptides.tsv", sep='\t', index=False)

detected_peptides.shape
