==> Jupyter Notebooks/4- Cal undetected peptides +apply SpC-NSAF.ipynb <==
# Calculate undetected peptides and apply SpC-NSAF

from google.colab import drive
drive.mount('/content/drive')

## Notebook summary:
1. Import fasta peptides and detected peptides datasets.
2. Dataset cleaning.
    - Sort detected peptides by PEP score.
    - Remove duplicates in both datasets.
    - Only keep proteins in fasta peptides that are also present in detected peptides - this is to be the expected peptides.
3. Get undetected peptides.
    - Remove all rows in expected peptides that are also present in detected peptides.
    - Calculate SpC-NSAF for undetected peptides
4. Export a new data frame with undetected peptides as a TSV file.

# import libraries
import numpy as np
import pandas as pd
import csv

# set display options
#pd.set_option("display.max_rows", None, "display.max_columns", None)

# 1. Import fasta peptides and detected peptides

# import datasets
fasta_peptides = pd.read_table('/content/drive/MyDrive/Colab Notebooks/peptide ml detection/data/fasta_peptides.tsv')
detected_peptides = pd.read_table('/content/drive/MyDrive/Colab Notebooks/peptide ml detection/data/detected_peptides_NSAF.tsv')

fasta_peptides.shape

detected_peptides.shape

# 2. Dataset cleaning

### get expected peptides before removing duplicates

# initial check on how many total proteins in detected_peptides are not in fasta_peptides (should be 0)
len(detected_peptides[~detected_peptides["Protein"].isin(fasta_peptides["Protein"])])

# initial check on how many total peptides (i.e. PSMs) in detected_peptides are not in fasta_peptides (should be 0)
len(detected_peptides[~detected_peptides["Peptide"].isin(fasta_peptides["Peptide"])])
# NB: These sequences are actually there, they appear to be start of the sequences
# i.e. all these sequences start with Methionine in the fasta peptides.

# initial check on how many total proteins and peptides in fasta_peptides are not in detected_peptides
len(fasta_peptides[~fasta_peptides["Protein"].isin(detected_peptides["Protein"])])

# check how many protein+peptide in detected_peptides are present in fasta_peptides (should be 37,847)
len(detected_peptides.set_index(['Protein', 'Peptide']).index.isin(fasta_peptides.set_index(['Protein', 'Peptide']).index))

len(fasta_peptides['Protein'])

## keep proteins in fasta peptides that are present in detected peptides -expected peptides

detected_peptides.describe(include='all')

# initial check on how many total proteins and peptides in detected_peptides are in fasta_peptides
# (should be 37,847 if after removing duplicate peptides)
len(detected_peptides[detected_peptides["Protein"].isin(fasta_peptides["Protein"])])

# initial check on how many total proteins in fasta_peptides are in detected_peptides
len(fasta_peptides[fasta_peptides["Protein"].isin(detected_peptides["Protein"])])

# remove proteins in fasta_peptides not present in detected_peptides - let this be expected_peptides
expected_peptides = fasta_peptides[fasta_peptides["Protein"].isin(detected_peptides["Protein"])]
expected_peptides.describe(include="all")

# 3. Get undetected peptides

# initial check on undetected peptides
undetected_peptides = pd.merge(expected_peptides, detected_peptides,
                               on=["Peptide"],
                               how='left', indicator=True)
undetected_peptides

# find peptides/PSMs present only in undetected_peptides
len(undetected_peptides[undetected_peptides['_merge'].str.contains("left_only", na=False)])

# find number of undetected using isin
len(expected_peptides[~expected_peptides["Peptide"].isin(detected_peptides["Peptide"])])

undetected_peptides = expected_peptides[~expected_peptides["Peptide"].isin(detected_peptides["Peptide"])]
undetected_peptides.shape

undetected_peptides.describe(include="all")

# remove any peptides containing 'U' (selenocysteine) amino acids
# this is so that we can map peptide amino acids correctly when calculating AAIndex1 properties
len(undetected_peptides[undetected_peptides['Peptide'].str.contains("U", na=False)])

undetected_peptides[undetected_peptides['Peptide'].str.contains("U", na=False)]

undetected_peptides.shape

len(undetected_peptides[~undetected_peptides['Peptide'].str.contains("U", na=False)])

undetected_peptides = undetected_peptides[~undetected_peptides['Peptide'].str.contains("U", na=False)]
undetected_peptides.shape

# Calculate SpC-NSAF for undetected peptides

# check for any intersection between detected and undetected (should be 0)
print(len(set(detected_peptides["Peptide"]).intersection(set(undetected_peptides["Peptide"]))))

# create dictionary from detected_peptides, with protein keys and associated NSAF values
detected_peptides_nsaf_dict = dict(zip(detected_peptides.Protein, detected_peptides.Quantification))

# map protein NSAF quantitation values to each protein in undetected_peptides
undetected_peptides['Quantification'] = undetected_peptides['Protein'].map(detected_peptides_nsaf_dict)
undetected_peptides

# check for any NaN values
undetected_peptides.isnull().sum()

print(undetected_peptides.shape)
print(detected_peptides.shape)

undetected_peptides.describe(include='all')

# remove any peptides in undetected_peptides that map to more than one different protein
undetected_peptides_clean = undetected_peptides.groupby('Peptide').filter(lambda x: x['Protein'].nunique() == 1)
undetected_peptides_clean.shape

undetected_peptides_clean.describe(include='all')

# drop duplicate peptides (i.e. if a peptide sequence that matches to same protein occurs twice or more), keep first occurrence
undetected_peptides_clean = undetected_peptides_clean.drop_duplicates(subset = ['Peptide', 'Protein'],
                                                                      keep = "first").reset_index(drop = True)
undetected_peptides_clean.describe(include="all")

# check how many detected proteins are not in undetected
len(detected_peptides[~detected_peptides["Protein"].isin(undetected_peptides_clean["Protein"])])

detected_peptides[~detected_peptides["Protein"].isin(undetected_peptides_clean["Protein"])]

# initial check on how many total proteins and PSMs in detected_peptides are in fasta_peptides
len(detected_peptides[detected_peptides["Protein"].isin(fasta_peptides["Protein"])])

undetected_peptides_clean.shape

# 4. Export a new data frame- undetected peptides as a TSV file

# export undetected peptides to TSV
undetected_peptides_clean.to_csv("/content/drive/MyDrive/Colab Notebooks/peptide ml detection/data/undetected_peptides_NSAF.tsv", sep='\t', index=False)
